refactor(redshift): Use atomic MERGE for execute_merge (#33)
Refactored the `RedshiftAdapter.execute_merge` method to use an atomic `MERGE` SQL statement instead of the previous non-atomic `DELETE` and `INSERT` logic.

This change improves the atomicity and performance of the merge operation, bringing it in line with Redshift best practices and the requirements of the FRD.

The corresponding integration test in `tests/test_redshift_adapter.py` has been updated to validate the new `MERGE` statement generation.