feat: Add BigQuery adapter and Slack alerter (#34)
This commit introduces two new features to the `py_load_pmda` package:

1.  A BigQuery adapter that allows users to load data into Google BigQuery.
2.  A Slack alerter that sends notifications to a Slack channel upon critical pipeline failure.

The BigQuery adapter implements the `LoaderInterface` and uses the `google-cloud-bigquery` library to interact with BigQuery. It supports bulk loading of data from pandas DataFrames and merging data using a staging table.

The Slack alerter implements the `Alerter` interface and uses the `slack_sdk` library to send messages to a Slack channel.

Both features are configurable through the `config.yaml` file and environment variables.

Tests have been added for both the BigQuery adapter and the Slack alerter.