Fix: Use snok/install-poetry action to fix CI build (#70)
The GitHub Actions workflow was failing because the poetry executable could not be found. This was because the `actions/setup-python` action with the `cache: 'poetry'` option no longer installs poetry.

This commit fixes the issue by using the `snok/install-poetry@v1` action to install poetry in the CI environment. This is the recommended way to install poetry in GitHub Actions.